requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
alembic==1.14.1
python-dotenv==1.0.1
celery==5.4.0
redis==5.2.1
openai==1.63.2
pydantic-settings==2.7.1
httpx==0.28.1
aiohttp==3.11.0
requests==2.31.0
websockets==14.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
email-validator==2.2.0
bcrypt==4.2.1
